requirements.txt
# Core Django
Django==5.1.4
mysqlclient==2.2.4

# Admin Theme
django-jazzmin==3.0.1

# Authentication & Authorization
django-allauth==65.3.0

# Forms & UI
django-crispy-forms==2.3
crispy-tailwind==1.0.3

# Image Processing
Pillow==11.0.0
django-imagekit==5.0.0

# API & REST
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Utilities
python-slugify==8.0.4
django-extensions==3.2.3

# Development
black==24.10.0
isort==5.13.2
pytest-django==4.9.0
